fix(recorder): delay task shutdown to finish reading stdout
Tasks may sometimes get stuck in the "Muxing" status. This happened
because when the process exits, a signal is sent to the threads which
are listening to the process stdout to shut down. Sometimes the signal
is sent before the process stdout could be read completely, which caused
the task to miss the final muxed filename.